correction multiple

+++ src/main/java/fr/eni/demo/dal/LocationRepository.java
@@ -4,8 +4,10 @@ import fr.eni.demo.bo.Location;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface LocationRepository extends MongoRepository<Location, Integer> {
+public interface LocationRepository extends MongoRepository<Location, Long> {
 
     Optional<Location> findByCodeBarre(String codeBarre);
